add ut, modify worning and so on.

+++ services/native/file_access_service/src/file_access_service.cpp
@@ -246,7 +246,7 @@ void FileAccessService::CleanRelativeObserver(const sptr<IFileAccessObserver> &o
         return obsContext->EqualTo(afterContext);
     });
     std::vector<Uri> uriLists = GetUriList(code);
-    for (int i = 0; i < uriLists.size(); ++i) {
+    for (size_t i = 0; i < uriLists.size(); ++i) {
         UnregisterNotify(uriLists[i], observer);
     }
 }
